+++ CHANGELOG.md
@@ -14,11 +14,13 @@ Changelog
 
 **Next release**
 
+[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
+
+**Version 1.99.58 TEST**
+
 * Fixed all minor bugs reported through the support info
 * Running [VACUUM](http://www.sqlite.org/lang_vacuum.html) on databases on boot
 * Setting [asynchronous mode](http://www.sqlite.org/pragma.html#pragma_synchronous) for usage data database
-
-[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 1.99.57 TEST**